requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.340
langchain-community==0.0.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pypdf==3.17.0
python-multipart==0.0.6
pydantic==2.5.0
rank-bm25==0.2.2
numpy==1.24.3
aiofiles==23.2.1
gradio==4.8.0
